verifily_cli_v1/requirements.txt
typer>=0.9.0
rich>=13.0.0
pyyaml>=6.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
httpx>=0.24.0
